Finish or so
I think it is ok like that. Even thought there seems to be difference with the picture,
my result do not look so bad and I find it functionnal like that. I am not sure
how I would reach the same result without tampering with the provided setting (quotation for one
-> I think it is too big but it does say 36px so ...).
I choose to make a picture even thought there is font&co requirements  because it say header logo, so ...
(and it was lacking cats even thought it was programmation).
I have some strange stuff that I am not clear how to deal with : I can not get the source for the quotation to do full right with the proportionnal sizing.

To do :
I forgot border for the pictures.